pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mkdocs-plugins"
version = "0.1.0"
description = "A collection of mkdocs plugins"
authors = [{ name="Erin Shaben", email="eshaben@icloud.com" }]
license = { text = "BSD-2-Clause" }
dependencies = [
  "mkdocs>=1.6",
  "mkdocs-material>=9.0",
  "pymdown-extensions>=10.0",
]

[project.entry-points."mkdocs.plugins"]
cards = "plugins.cards.plugin:CardsPlugin"

# Configuration for development tools
[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88